fix: 실시간 랭킹 구현 완료 (Redis 기반)

+++ backend/src/main/java/com/newlearn/backend/rank/model/UserRank.java
@@ -31,7 +31,7 @@ public class UserRank {
     private String rankingType;
 
     @Column(name = "ranking", nullable = false)
-    private int ranking;
+    private Long ranking;
 
     @Column(name = "score", nullable = false)
     private Long score;
